+++ src/xdoc/error.rs
@@ -14,7 +14,7 @@ pub struct XDocErr {
     /// The sourcecode line where the error was raised.
     pub line: u64,
     /// The underlying error that is being wrapped.
-    pub source: Option<Box<dyn Error>>,
+    pub source: Option<Box<dyn Error + Sync + Send + 'static>>,
 }
 
 impl fmt::Display for XDocErr {
@@ -41,52 +41,4 @@ impl Error for XDocErr {
         }
         None
     }
-}
-
-macro_rules! wrap {
-    // Base case:
-    ($err:expr) => (Err($crate::xdoc::error::XDocErr {
-        message: "an error occurred".to_string(),
-        file: file!().to_string(),
-        line: line!() as u64,
-        source: Some($err.into()),
-    }));
-    ($err:expr, $msg:expr) => (Err($crate::xdoc::error::XDocErr {
-        message: $msg.to_string(),
-        file: file!().to_string(),
-        line: line!() as u64,
-        source: Some($err.into()),
-    }));
-    ($err:expr, $fmt:expr, $($arg:expr),+) => (Err($crate::xdoc::error::XDocErr {
-        message: format!($fmt, $($arg),+),
-        file: file!().to_string(),
-        line: line!() as u64,
-        source: Some($err.into()),
-    }));
-}
-
-macro_rules! better_wrap {
-    ($result:expr) => {
-        match $result {
-            Ok(value) => Ok(value),
-            Err(e) => wrap!(e),
-        }
-    };
-}
-
-// a convenience macro for creating a Result::Err
-macro_rules! raise {
-    // Base case:
-    ($msg:expr) => (Err($crate::xdoc::error::XDocErr {
-        message: $msg.to_string(),
-        file: file!().to_string(),
-        line: line!() as u64,
-        source: None,
-    }));
-    ($fmt:expr, $($arg:expr),+) => ($Err($crate::xdoc::error::XErr {
-        message: format!($fmt, $($arg),+),
-        file: file!().to_string(),
-        line: line!() as u64,
-        source: None,
-    }));
 }