+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Currently we are using v0.0.x where every version can and will contain breaking 
 ## [Unreleased]
 ## Added
 - Whitespace normalization of text nodes [#75]
+- Support CDATA sections [#76]
 
 ### Changed
 - Make `Element` struct members private [#74]
@@ -16,6 +17,7 @@ Currently we are using v0.0.x where every version can and will contain breaking 
 
 [#74]: https://github.com/webern/exile/pull/74
 [#75]: https://github.com/webern/exile/pull/75
+[#76]: https://github.com/webern/exile/pull/76
 
 ## [v0.0.2] - 2020-11-15
 ### Added